removed try catch blocks. added areaname and battleground scraping to field archive. trying to figure out stack args for rbg shadeloop
