+++ Core/Field/JSM/Instructions/BGSHADE.cs
@@ -1,44 +1,41 @@
-﻿using System;
+﻿using Microsoft.Xna.Framework;
+using System;
 
 
 namespace OpenVIII.Fields.Scripts.Instructions
 {
-    internal sealed class BGSHADE : JsmInstruction
+    /// <summary>
+    /// Shade between two colors
+    /// </summary>
+    /// <see cref="http://wiki.ffrtt.ru/index.php?title=FF8/Field/Script/Opcodes/0D0_BGSHADE&action=edit&redlink=1"/>
+    public sealed class BGSHADE : JsmInstruction
     {
-        private IJsmExpression _arg0;
-        private IJsmExpression _arg1;
-        private IJsmExpression _arg2;
-        private IJsmExpression _arg3;
-        private IJsmExpression _arg4;
-        private IJsmExpression _arg5;
-        private IJsmExpression _arg6;
+        private readonly Int32 _arg0;
+        private readonly Color _c0;
+        private readonly Color _c1;
 
-        public BGSHADE(IJsmExpression arg0, IJsmExpression arg1, IJsmExpression arg2, IJsmExpression arg3, IJsmExpression arg4, IJsmExpression arg5, IJsmExpression arg6)
+        public BGSHADE(Int32 arg0, byte red0, byte green0, byte blue0, byte red1, byte green1, byte blue1)
         {
-            _arg0 = arg0;
-            _arg1 = arg1;
-            _arg2 = arg2;
-            _arg3 = arg3;
-            _arg4 = arg4;
-            _arg5 = arg5;
-            _arg6 = arg6;
+            _arg0 = arg0; //unknown
+            (_c0.R, _c0.G, _c0.B, _c0.A) = (red0, green0, blue0, 0xFF); //red and blue could be reversed.
+            (_c1.R, _c1.G, _c1.B, _c1.A) = (red1, green1, blue1, 0xFF); //red and blue could be reversed.
         }
 
         public BGSHADE(Int32 parameter, IStack<IJsmExpression> stack)
             : this(
-                arg6: stack.Pop(),
-                arg5: stack.Pop(),
-                arg4: stack.Pop(),
-                arg3: stack.Pop(),
-                arg2: stack.Pop(),
-                arg1: stack.Pop(),
-                arg0: stack.Pop())
+                blue1: ((IConstExpression)stack.Pop()).Byte(),
+                green1: ((IConstExpression)stack.Pop()).Byte(),
+                red1: ((IConstExpression)stack.Pop()).Byte(),
+                blue0: ((IConstExpression)stack.Pop()).Byte(),
+                green0: ((IConstExpression)stack.Pop()).Byte(),
+                red0: ((IConstExpression)stack.Pop()).Byte(),
+                arg0: ((IConstExpression)stack.Pop()).Int32())
         {
         }
 
         public override String ToString()
         {
-            return $"{nameof(BGSHADE)}({nameof(_arg0)}: {_arg0}, {nameof(_arg1)}: {_arg1}, {nameof(_arg2)}: {_arg2}, {nameof(_arg3)}: {_arg3}, {nameof(_arg4)}: {_arg4}, {nameof(_arg5)}: {_arg5}, {nameof(_arg6)}: {_arg6})";
+            return $"{nameof(BGSHADE)}({nameof(_arg0)}: {_arg0}, {nameof(_c0)}: {_c0}, {nameof(_c1)}: {_c1})";
         }
     }
 }
