+++ Core/Field/JSM/Instructions/RBGSHADELOOP.cs
@@ -1,53 +1,62 @@
-﻿using System;
+﻿using Microsoft.Xna.Framework;
+using System;
 
 
 namespace OpenVIII.Fields.Scripts.Instructions
 {
-    internal sealed class RBGSHADELOOP : JsmInstruction
+    /// <summary>
+    /// RBG Shade Loop between two colors.
+    /// </summary>
+    /// <see cref="http://wiki.ffrtt.ru/index.php?title=FF8/Field/Script/Opcodes/0D2_RBGSHADELOOP&action=edit&redlink=1"/>
+    public sealed class RBGSHADELOOP : JsmInstruction
     {
-        private IJsmExpression _arg0;
-        private IJsmExpression _arg1;
-        private IJsmExpression _arg2;
-        private IJsmExpression _arg3;
-        private IJsmExpression _arg4;
-        private IJsmExpression _arg5;
-        private IJsmExpression _arg6;
-        private IJsmExpression _arg7;
-        private IJsmExpression _arg8;
-        private IJsmExpression _arg9;
+        private readonly Int32 _frames0;
+        private readonly Int32 _frames1;
+        private readonly Int32 _arg8;
+        private readonly Int32 _arg9;
+        private readonly Color _c0;
+        private readonly Color _c1;
 
-        public RBGSHADELOOP(IJsmExpression arg0, IJsmExpression arg1, IJsmExpression arg2, IJsmExpression arg3, IJsmExpression arg4, IJsmExpression arg5, IJsmExpression arg6, IJsmExpression arg7, IJsmExpression arg8, IJsmExpression arg9)
+        public int Arg0 => _frames0;
+
+        public int Arg1 => _frames1;
+
+        public int Arg8 => _arg8;
+
+        public int Arg9 => _arg9;
+
+        public Color C0 => _c0;
+
+        public Color C1 => _c1;
+
+        public RBGSHADELOOP(Int32 frames0, Int32 frames1, byte red0, byte green0, byte blue0, byte red1, byte green1, byte blue1, Int32 arg8, Int32 arg9)
         {
-            _arg0 = arg0;
-            _arg1 = arg1;
-            _arg2 = arg2;
-            _arg3 = arg3;
-            _arg4 = arg4;
-            _arg5 = arg5;
-            _arg6 = arg6;
-            _arg7 = arg7;
-            _arg8 = arg8;
-            _arg9 = arg9;
+            _frames0 = frames0; // transitional frame time
+            _frames1 = frames1; // transitional frame time
+            (_c0.R, _c0.G, _c0.B,_c0.A) = (red0, green0, blue0, 0xFF); 
+            (_c1.R, _c1.G, _c1.B, _c1.A) = (red1, green1, blue1, 0xFF);
+            _arg8 = arg8; //donno
+            _arg9 = arg9; //donno
         }
 
         public RBGSHADELOOP(Int32 parameter, IStack<IJsmExpression> stack)
             : this(
-                arg9: stack.Pop(),
-                arg8: stack.Pop(),
-                arg7: stack.Pop(),
-                arg6: stack.Pop(),
-                arg5: stack.Pop(),
-                arg4: stack.Pop(),
-                arg3: stack.Pop(),
-                arg2: stack.Pop(),
-                arg1: stack.Pop(),
-                arg0: stack.Pop())
+                arg9: ((IConstExpression)stack.Pop()).Int32(),
+                arg8: ((IConstExpression)stack.Pop()).Int32(),
+                blue1: ((IConstExpression)stack.Pop()).Byte(),
+                green1: ((IConstExpression)stack.Pop()).Byte(),
+                red1: ((IConstExpression)stack.Pop()).Byte(),
+                blue0: ((IConstExpression)stack.Pop()).Byte(),
+                green0: ((IConstExpression)stack.Pop()).Byte(),
+                red0: ((IConstExpression)stack.Pop()).Byte(),
+                frames1: ((IConstExpression)stack.Pop()).Int32(),
+                frames0: ((IConstExpression)stack.Pop()).Int32())
         {
         }
 
         public override String ToString()
         {
-            return $"{nameof(RBGSHADELOOP)}({nameof(_arg0)}: {_arg0}, {nameof(_arg1)}: {_arg1}, {nameof(_arg2)}: {_arg2}, {nameof(_arg3)}: {_arg3}, {nameof(_arg4)}: {_arg4}, {nameof(_arg5)}: {_arg5}, {nameof(_arg6)}: {_arg6}, {nameof(_arg7)}: {_arg7}, {nameof(_arg8)}: {_arg8}, {nameof(_arg9)}: {_arg9})";
+            return $"{nameof(RBGSHADELOOP)}({nameof(_frames0)}: {_frames0}, {nameof(_frames1)}: {_frames1}, {nameof(_c0)}: {_c0}, {nameof(_c1)}: {_c1}, {nameof(_arg8)}: {_arg8}, {nameof(_arg9)}: {_arg9})";
         }
     }
 }
