+++ DirectX/Program.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Diagnostics;
 
 namespace OpenVIII.DirectX
 {
@@ -16,17 +15,8 @@ namespace OpenVIII.DirectX
         [STAThread]
         private static void Main()
         {
-            start:
-            try
-            {
-                using (Game1 game = new Game1())
-                    game.Run();
-            }
-            catch (InvalidOperationException e)
-            {
-                Debug.WriteLine(e);
-                goto start;
-            }
+            using (Game1 game = new Game1())
+                game.Run();
         }
     }
